chore: Reorganize spin-oci modules
